Added autostart of startx

+++ dwm/install.sh
@@ -15,9 +15,12 @@ main() {
             'i') su -c "cp status.sh /usr/bin/dwmstatus && chmod +rwx /usr/bin/dwmstatus" root
                  make
                  su -c "make clean install" root
-                 mv ~/.xinitrc ~/.xinitrc-old
+                 if [ -e ~/.xinitrc ]; then
+                    mv ~/.xinitrc ~/.xinitrc-old
+                 fi
                  echo "picom & dwmstatus & exec dwm" > ~/.xinitrc
                  chmod +rwx .xinitrc
+                 echo "if [[ -z $DISPLAY ]] && [[ $(tty) = /dev/tty1 ]]; then startx fi" > ~/.bash_profile
                  echo "Install complete! Reboot for changes to take effect" ;;
             'q') echo "Quitting..."
                  exit 0 ;;
